+++ CHANGELOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - edit validation of db names to give the possibility to use name with underscore
 - remove transaction for catalog_product_flat reindexAll
 - add config to toggle the transaction for catalog_product_flat
+- add configurable cron for reindex